Done giao dien va tac vu them Rating & Review

+++ app/src/main/java/com/nhathao/e_commerce/models/Rating.kt
@@ -4,6 +4,6 @@ class Rating (
     var ratingId: String,
     var userAccountName: String,
     var productId:String,
-    var ratingStar:Int
+    var ratingStar:Float
 ){
 }
